initial server.ts changes and adding types.ts
