+++ server/types.ts
@@ -1,0 +1,5 @@
+export type ServerError = {
+    log: string;
+    message: {err: string};
+    status: number;
+  }
